+++ src/components/layouts/Announcements.js
