+++ src/components/layouts/joblogger/JobLoggerStateIndicator.js
