+++ src/components/pages/manage/productions/create/Event.js
@@ -13,7 +13,7 @@ import { Text } from '@/ui/Text';
 import { Title } from '@/ui/Title';
 
 import { useGetCalendarSummary } from '@/hooks/api/events';
-import { CalendarType } from '../../../../constants/CalendarType';
+import { CalendarType } from '@/constants/CalendarType';
 
 const Event = ({
   terms,
