+++ src/components/pages/place/[placeId]/status/StatusModal.js
@@ -1,11 +1,10 @@
 import PropTypes from 'prop-types';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { MaxLengthReason } from './constants';
 import { Modal, ModalVariants } from '@/ui/Modal';
 import { Stack } from '@/ui/Stack';
-import { ReasonAndTypeForm } from './ReasonAndTypeForm';
-import { OfferType } from '../../constants/OfferType';
+import { ReasonAndTypeForm } from '@/pages/ReasonAndTypeForm';
+import { OfferType } from '@/constants/OfferType';
 
 const StatusModal = ({ visible, className, onClose }) => {
   const { t } = useTranslation();
@@ -24,7 +23,7 @@ const StatusModal = ({ visible, className, onClose }) => {
       cancelTitle={t('offerStatus.actions.close')}
       onConfirm={() => {}}
       onClose={onClose}
-      confirmButtonDisabled={reason.length > MaxLengthReason}
+      confirmButtonDisabled={reason.length > 200}
     >
       <Stack padding={4}>
         <ReasonAndTypeForm
