+++ src/hooks/api/authenticated-query.test.js
@@ -4,8 +4,8 @@ import {
   useAuthenticatedQuery,
 } from './authenticated-query';
 
-import { mockResponses, setupPage } from '../../tests/utils/setupPage';
-import { renderHookWithWrapper } from '../../tests/utils/renderHookWithWrapper';
+import { mockResponses, setupPage } from '../../test/utils/setupPage';
+import { renderHookWithWrapper } from '../../test/utils/renderHookWithWrapper';
 import { fetchFromApi } from '@/utils/fetchFromApi';
 
 const queryFn = async ({ headers, ...queryData }) => {
