--- src/tests/mocks/mockRouterWithParams.js
@@ -1,20 +1,0 @@
-const mockRouterWithParams = ({ query, ...rest } = {}) => {
-  const useRouter = jest.spyOn(require('next/router'), 'useRouter');
-
-  const push = jest.fn();
-
-  const mockRouter = {
-    pathname: '/',
-    query: {
-      ...(query ?? {}),
-    },
-    asPath: '/',
-    push,
-    ...rest,
-  };
-
-  useRouter.mockImplementation(() => mockRouter);
-  return mockRouter;
-};
-
-export { mockRouterWithParams };
